[ascl] Refine preprocessor     -   removed imps from ast_t     -   removed ast_imp.[ch]     -   changed to preprocess before parsing     -   renamed ERROR_INFINITE_IMPORT to ERROR_CYCLIC_IMPORT

+++ contract/native/src_pos.c
@@ -34,7 +34,7 @@ src_pos_print(src_pos_t *pos, char *buf, int buf_size)
         adj_offset += pos->first_col - adj_col;
     }
 
-    adj_len = MIN(strlen(pos->src + adj_offset), line_max);
+    adj_len = MIN((int)strlen(pos->src + adj_offset), line_max);
 
     for (i = 0; i < adj_len; i++) {
         char c = pos->src[i + adj_offset];
